requirements.txt
flask>=2.0.1
pytest>=7.0.0
bandit>=1.7.0
safety>=2.3.5
flask-sqlalchemy>=2.5.1
flask-login>=0.5.0
flask-wtf>=0.15.1
flask-bcrypt>=0.7.1
python-dotenv>=0.19.0
email-validator>=1.1.3
Pillow>=9.0.0
flask-limiter>=2.4.0
flask-talisman>=0.8.1
flask-helmet>=0.1.0
flask-cors>=3.0.10 